% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_lipids}
\alias{annotate_lipids}
\title{Parse Molecule names to extract class and chain information.}
\usage{
annotate_lipids(molecules)
}
\arguments{
\item{molecules}{A character vector containing lipid molecule names}
}
\value{
a data.frame with lipid annotations as columns. Input lipid names
are given in a column named "Molecule"
}
\description{
Parse Molecule names to extract class and chain information.
}
