% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{as_skyline_experiment}
\alias{as_skyline_experiment}
\title{Convert data.frame/matrix to LipidomicsExperiment}
\usage{
as_skyline_experiment(df, logged = FALSE, normalized = FALSE)
}
\arguments{
\item{df}{A data.frame or matrix where rows are lipids and columns
are samples. Lipid names should be provided in the first column
or in rownames of \code{df}. Measurements should be numeric.
The data is considered \code{summarized} unless at least one lipid
is duplicated (has > 1 row).}

\item{logged}{Whether the data is log-transformed}

\item{normalized}{Whether the data is normalized}
}
\value{
LipidomicsExperiment
}
\description{
Convert data.frame/matrix to LipidomicsExperiment
}
