% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidr-package.R
\docType{data}
\name{data_nomalized}
\alias{data_nomalized}
\title{Example dataset (normalized and log transformed)}
\description{
Please see \link{normalize_pqn} for details on how to generate this dataset.
}
\examples{
data(data_nomalized)
}
