% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{de_design}
\alias{de_design}
\title{Differential analysis of lipids between sample groups}
\usage{
de_design(..., coef = NULL, design, data, measure = "Area")
}
\arguments{
\item{...}{expressions, or character strings which can be parsed to expressions, specifying contrasts.
These are passed to \code{limma::makeContrasts}. Ignored if \code{coef} is provided.}

\item{coef}{column number or column name specifying which coefficient of the linear model is of interest.}

\item{design}{Design matrix generated from \code{\link{model.matrix}}, or a design formula .}

\item{data}{Skyline data.frame created by \code{\link{read_skyline}}, should be normalized and log2 transformed}

\item{measure}{name of the column indicating sample names}
}
\description{
Differential analysis of lipids between sample groups
}
\examples{
# type ?normalize_pqn to see how to normalize and log-transfome your data
data(data_normalized)

# Using formula
de_results = de_design(coef="groupHighFat_water", design = ~ group, data=d_normalized, measure="Area")

model.matrix(~ group, data=colData(d_normalized))
de_results = de_design(coef="groupHighFat_water", design = design, data=d_normalized, measure="Area")
}
