% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{.as_regex}
\alias{.as_regex}
\title{Regex-escaping for character vectors.}
\usage{
.as_regex(strings, collapse = FALSE, prefix = "", suffix = "")
}
\arguments{
\item{strings}{A character vector to be regex-escaped.}

\item{collapse}{Collapse all strings to create a single pattern (using \code{|}).}
}
\value{
regex-escaped string to be used for pattern matching
}
\description{
Regex-escaping for character vectors.
}
