% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfiles.R
\name{.read_skyline_file}
\alias{.read_skyline_file}
\title{Internal method to read skyline file}
\usage{
.read_skyline_file(file)
}
\arguments{
\item{file}{skyline exported file in CSV format}
}
\value{
std data.frame
}
\description{
Internal method to read skyline file
}
