% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{enrich_lipidsets}
\alias{enrich_lipidsets}
\title{Lipid set enrichment analysis}
\usage{
enrich_lipidsets(de.results, rank.by = c("logFC", "P.Value",
  "Adj.P.Val"))
}
\arguments{
\item{de.results}{output of \code{\link{de_analysis}}}

\item{rank.by}{statistic used to rank the lipid list}
}
\description{
Lipid set enrichment analysis
}
