% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw.R
\name{fetch_mw_study}
\alias{fetch_mw_study}
\title{Download and parse full data for a study from Metabolomics Workbench.}
\usage{
fetch_mw_study(study_id)
}
\arguments{
\item{study_id}{The Metabolomics Workbench study ID.}
}
\value{
A LipidomicsExperiment object containing clinical and lipid intensity
data.
}
\description{
This functions uses Metabolomics Workbench REST API to retrieve study
data using a study ID. The function returns a LipidomicsExperiment where users
can directly apply \code{lipidr} analysis workflow.
}
\examples{
# fetch_mw_study("ST001111")
}
\seealso{
Other Metabolomics Workbench: \code{\link{list_mw_studies}},
  \code{\link{read_mwTab}},
  \code{\link{read_mw_datamatrix}}
}
\concept{Metabolomics Workbench}
