% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{filter_by_blanks}
\alias{filter_by_blanks}
\title{Used to filter out molecules that read below a density threshold.}
\usage{
filter_by_blanks(data, fold.cutoff = 10, measure = "Area")
}
\arguments{
\item{data}{LipidomicsExperiment object.}

\item{fold.cutoff}{Fold Chain threshold (numeric). Default is '10'.}

\item{measure}{Which is used to measure fold chain, usually log2(Area) (default).}
}
\value{
LipidomicsExperiment object with molecules filtered
}
\description{
Used to filter out molecules that read below a density threshold.
}
\examples{
data(data_normalized)
filter_by_blanks(data_normalized)
}
