% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{filter_by_cv}
\alias{filter_by_cv}
\title{Remove molecules with CV larger that a threshold}
\usage{
filter_by_cv(data, cv.cutoff = 20, measure = "Area")
}
\arguments{
\item{data}{LipidomicsExperiment object created by \code{\link[=read_skyline]{read_skyline()}}.}

\item{cv.cutoff}{CV threshold (numeric).  Default is \code{20}.}

\item{measure}{Which measure used to calculate CV, usually Area (default).}
}
\value{
LipidomicsExperiment object with molecules filtered.
}
\description{
Remove molecules with CV larger that a threshold
}
\examples{
data(data_normalized)
filter_by_cv(data_normalized)
}
