% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsea.R
\name{gen_lipidsets}
\alias{gen_lipidsets}
\title{Generate lipid sets from lipid molecule names}
\usage{
gen_lipidsets(molecules, min_size = 2)
}
\arguments{
\item{molecules}{A character vector containing lipid molecule names.}

\item{min_size}{Minimum number of molecules in a set to be included
in enrichment.}
}
\value{
List of lipid sets
}
\description{
Generate lipid sets from lipid molecule names
}
\examples{
data(data_normalized)
molecules <- rowData(data_normalized)$Molecule
gen_lipidsets(molecules)
}
