% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidr-package.R
\docType{data}
\name{lipidnames_pattern}
\alias{lipidnames_pattern}
\title{Patterns used in parsing lipid names}
\description{
A collection of patterns to extract lipid class and chain information
from lipid names. Used internally by the package.
}
\examples{
data(lipidnames_pattern)
}
