% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidr-package.R
\docType{data}
\name{lipidr-data}
\alias{lipidr-data}
\title{Description of lipidr datasets}
\description{
lipidr-package has 3 datasets: \itemize{
\item data_normalized    Example lipidomics dataset,
normalized & log2-transformed.
\item lipidDefaults   A list of default mappings and annotations for lipids.
\item lipidname_pattern   A list of patterns used in parsing lipid names.
}
See below for detailed description of each dataset.
}
\examples{
data(data_normalized)
}
\seealso{
Other lipidr datasets: \code{\link{data_normalized}},
  \code{\link{lipidDefaults}},
  \code{\link{lipidnames_pattern}}
}
\concept{lipidr datasets}
