% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw.R
\name{list_mw_studies}
\alias{list_mw_studies}
\title{Retrieve list of lipidomics studies from Metabolomics Workbench.}
\usage{
list_mw_studies(keyword = "lipid")
}
\arguments{
\item{keyword}{A keyword to search for in Metabolomics Workbench studies.}
}
\value{
A data frame with studies matching the keyword. Study ID, title,
author and details are retrieved.
}
\description{
This functions uses Metabolomics Workbench REST API to retrieve summaries
of lipidomics studies.
}
\examples{
# list_mw_studies()
}
\seealso{
Other Metabolomics Workbench: \code{\link{fetch_mw_study}},
  \code{\link{read_mwTab}},
  \code{\link{read_mw_datamatrix}}
}
\concept{Metabolomics Workbench}
