% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{non_parsed_molecules}
\alias{non_parsed_molecules}
\title{Get a list of molecules that couldn't be parsed by \code{lipidr}}
\usage{
non_parsed_molecules(data)
}
\arguments{
\item{data}{LipidomicsExperiment object created by \code{\link[=read_skyline]{read_skyline()}}.}
}
\value{
A character vector of the molecule names that could not be parsed.
}
\description{
Get a list of molecules that couldn't be parsed by \code{lipidr}
}
\examples{
data(data_normalized)
non_parsed_molecules(data_normalized)
}
