% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_itsd}
\alias{normalize_itsd}
\title{Normalize each class by its corresponding internal standard(s).}
\usage{
normalize_itsd(data, measure = "Area", exclude = "blank", log = TRUE)
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link{read_skyline}}}

\item{measure}{which measure to use as intensity, usually Area, Area.Normalized or Height}

\item{exclude}{Samples to exclude, can be either: \cr 
"blank" - automatically detected blank samples and exclude them
logical vector with the same length as samples}

\item{log}{whether the normalized values should be log2 transformed}
}
\value{

}
\description{
Normalize each class by its corresponding internal standard(s).
Lipid classes are normalized using corresponding internal standard(s)
of the same lipid class. If no corresponding internal standard is found
the average of all measured internal standards is used instead.
}
