% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lipidclass}
\alias{plot_lipidclass}
\title{Inofrmative plots to investigate lipid classes}
\usage{
plot_lipidclass(data, type = c("sd", "boxplot"), measure = "Area",
  log = TRUE)
}
\arguments{
\item{data}{SkylineExperiment object created by \code{\link[=read_skyline]{read_skyline()}}.}

\item{type}{plot type, either \code{sd} or \code{boxplot}. Default is \code{sd}}

\item{measure}{Which measure to plot the distribution of: usually Area,
Area.Normalized, Height or Retention.Time}

\item{log}{Whether values should be log2 transformed
(Set FALSE for retention time).}
}
\value{
A ggplot object.
}
\description{
\code{lipidr} supports two types of plots for to look at lipid classes.
\code{sd} plots a bar chart for standard deviation of a certain measure in each
class. This plot type is usually used to look at standard deviations of
intensity in each class, but can also be used to look at different measures
such as \code{Retention.Time}, to ensure all lipids are eluted within the expected
range. To assess instrumental variation apply the function to technical
quality control samples.
\code{boxplot} Plots a boxplot chart to examine the distribution of values per
class. This plot type is usually used to look at the intensity distribution
in eachclass, but can also be used to look at different measures, such as
\code{Retention.Time} or \code{Background}.
}
\examples{
data(data_normalized)

d_qc <- data_normalized[, data_normalized$group == "QC"]
plot_lipidclass(d_qc, "sd", "Area", log = TRUE)
plot_lipidclass(d_qc, "sd", "Retention.Time", log = FALSE)
plot_lipidclass(d_qc, "boxplot", "Area", log = TRUE)
plot_lipidclass(d_qc, "boxplot", "Retention.Time", log = FALSE)
}
