% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_molecule_boxplot}
\alias{plot_molecule_boxplot}
\title{Plot a boxplot chart to examine the distribution of values per lipid}
\usage{
plot_molecule_boxplot(data, measure = "Area", log = TRUE)
}
\arguments{
\item{data}{SkylineExperiment object created by \code{\link[=read_skyline]{read_skyline()}}.}

\item{measure}{Which measure to plot the distribution of: usually Area,
Area.Normalized or Height.}

\item{log}{Whether values should be log2 transformed.}
}
\value{
A ggplot object.
}
\description{
The function is usually used to look at intensity distribution for each
lipid, but can also be used to look at different measures, such as
\code{Retention.Time} or \code{Background}.
}
\examples{
data(data_normalized)

# plot the variation in intensity of ITSD (internal standards) in QC samples
d_itsd_qc <- data_normalized[
  rowData(data_normalized)$itsd,
  data_normalized$group == "QC"
]
plot_molecule_boxplot(d_itsd_qc)
plot_molecule_boxplot(d_itsd_qc, "Retention.Time", log = FALSE)
}
