% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_molecule_sd}
\alias{plot_molecule_sd}
\title{Plot a bar chart for standard deviation if a certain measure in each lipid}
\usage{
plot_molecule_sd(data, measure = "Area", log = TRUE)
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link{read_skyline}}}

\item{measure}{Which measure to plot the distribution of: usually Area, Area.Normalized or Height}

\item{log}{Whether values should be log2 transformed (Set FALSE for retention time).}
}
\description{
The function is usually used to look at standard deviation of intensity for each lipid,
but can also be used to look at different measures such as  `Retention.Time`,
to ensure all lipids elute within expected range.
}
\examples{
datadir = system.file("extdata", package="lipidr")
filelist = list.files(datadir, "data.csv", full.names = TRUE)
d = read_skyline(filelist)

plot_molecule_sd(d)
plot_molecule_sd(d, "Retention.Time", log=FALSE)
}
