% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sample_boxplot}
\alias{plot_sample_boxplot}
\title{Plot a boxplot chart to examine the distribution of values per sample}
\usage{
plot_sample_boxplot(data, measure = "Area", log = TRUE)
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link[=read_skyline]{read_skyline()}}.}

\item{measure}{Which measure to plot the distribution of: usually Area,
Area.Normalized or Height.}

\item{log}{Whether values should be log2 transformed.}
}
\value{
A ggplot object.
}
\description{
The function should usually be used to look at intensity distribution in
each sample ensuring they are normalized. It can also be used to look at
different measures such as  \code{Retention.Time} or \code{Background}.
}
\examples{
data(data_normalized)

plot_sample_boxplot(data_normalized, "Area", log = TRUE)
plot_sample_boxplot(
  data_normalized[, data_normalized$group == "QC"],
  measure = "Retention.Time", log = FALSE
)
}
