% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sample_tic}
\alias{plot_sample_tic}
\title{Plot a bar chart for total sample intensity}
\usage{
plot_sample_tic(data, measure = "Area", log = TRUE)
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link[=read_skyline]{read_skyline()}}.}

\item{measure}{Which measure to use as intensity, usually Area, Area.Normalized or Height.}

\item{log}{Whether values should be log2 transformed.}
}
\value{
A ggplot object.
}
\description{
Plot a bar chart for total sample intensity
}
\examples{
data(data_normalized)

plot_sample_tic(data_normalized, "Area", log = TRUE)
plot_sample_tic(data_normalized, "Background", log = FALSE)
}
