% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw.R
\name{read_mw_datamatrix}
\alias{read_mw_datamatrix}
\title{Parse a Metabolomics Workbench data matrix into a SkylineExperiment.
Data matrix downloaded from Metabolomics Workbench are parsed into
a SkylineExperiment object to enable \code{lipidr} workflow analysis.}
\usage{
read_mw_datamatrix(file)
}
\arguments{
\item{file}{File path or url for the file containing the data matrix.}
}
\value{
A SkylineExperiment object containing clinical and lipid intensity
data.
}
\description{
Parse a Metabolomics Workbench data matrix into a SkylineExperiment.
Data matrix downloaded from Metabolomics Workbench are parsed into
a SkylineExperiment object to enable \code{lipidr} workflow analysis.
}
\seealso{
Other Metabolomics Workbench: \code{\link{fetch_mw_study}},
  \code{\link{list_mw_studies}}, \code{\link{read_mwTab}}
}
\concept{Metabolomics Workbench}
