% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw.R
\name{read_mwTab}
\alias{read_mwTab}
\title{Parse mwTab file into a SkylineExperiment.}
\usage{
read_mwTab(mwTab)
}
\arguments{
\item{file}{File path or url for a mwTab file.}
}
\value{
A SkylineExperiment object containing clinical and lipid intensity
data.
}
\description{
Parse mwTab file into a SkylineExperiment.
}
\seealso{
Other Metabolomics Workbench: \code{\link{fetch_mw_study}},
  \code{\link{list_mw_studies}},
  \code{\link{read_mw_datamatrix}}
}
\concept{Metabolomics Workbench}
