% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{is_logged}
\alias{is_logged}
\alias{set_logged}
\alias{is_normalized}
\alias{set_normalized}
\alias{is_summarized}
\alias{set_summarized}
\title{Functions to get and set attributes of SkylineExperiment objects}
\usage{
is_logged(data, measure)

set_logged(data, measure, val)

is_normalized(data, measure)

set_normalized(data, measure, val)

is_summarized(data)

set_summarized(data, val)
}
\arguments{
\item{data}{SkylineExperiment object created by \code{\link[=read_skyline]{read_skyline()}}.}

\item{measure}{Which measure to get / set attributes of.}

\item{val}{Value to be assigned to the attribute.}
}
\value{
Modified SkylineExperiment.
}
\description{
Functions to get and set attributes of SkylineExperiment objects
}
\examples{
data(data_normalized)
is_logged(data_normalized, "Area")
is_summarized(data_normalized)
}
