% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{significant_molecules}
\alias{significant_molecules}
\title{Get a list of significantly changed molecules}
\usage{
significant_molecules(de.results, p.cutoff = 0.05, logFC.cutoff = 1)
}
\arguments{
\item{de.results}{output of \code{\link{de_analysis}}}

\item{p.cutoff}{Sigificance threshold}

\item{logFC.cutoff}{Cut off for log Fold change}
}
\value{
a character vector with names of singificantly differentially changed lipids
}
\description{
Get a list of significantly changed molecules
}
\examples{
data(data_normalized)
de_results = de_analysis(HighFat_water - NormalDiet_water, data=data_normalized, measure="Area")
significant_molecules(de_results)
}
