% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\docType{class}
\name{SkylineExperiment-class}
\alias{SkylineExperiment-class}
\alias{.SkylineExperiment}
\title{SkylineExperiment object}
\description{
SkylineExperiment object
}
\section{Slots}{

\describe{
\item{\code{attrs}}{Extra slot to hold the workflow stage for the data,
whether normalized, summarized or log transformed.}
}}

