% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{SkylineExperiment}
\alias{SkylineExperiment}
\title{Constructor for Skyline experiment from list of Assays}
\usage{
SkylineExperiment(assay_list, attrs, colData = NULL, rowData = NULL,
  ...)
}
\description{
Constructor for Skyline experiment from list of Assays
}
