% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_transitions}
\alias{summarize_transitions}
\title{Collapse transitions}
\usage{
summarize_transitions(data, method = c("max", "average"))
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link{read_skyline}}}

\item{method}{choose to summarize multiple transitions by taking average or max intensity}
}
\description{
Calculate a single intensity for molecules with multiple transitions
}
