% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{update_molecule_names}
\alias{update_molecule_names}
\title{Rename molecules in a dataset
This function enables users to rename selected molecules in the dataset,
so that they can be parsed correctly by \code{lipidr} or modify the lipidclass.
\code{lipidr} automatically updates the annotation for the renamed molecules.}
\usage{
update_molecule_names(data, old, new)
}
\arguments{
\item{data}{SkylineExperiment object created by \code{\link[=read_skyline]{read_skyline()}}.}

\item{old}{A character vector of the molecule names to be renamed.}

\item{new}{A character vector of the new molecule names.}
}
\value{
A SkylineExperiment object with molecules name and annotation
updated.
}
\description{
Rename molecules in a dataset
This function enables users to rename selected molecules in the dataset,
so that they can be parsed correctly by \code{lipidr} or modify the lipidclass.
\code{lipidr} automatically updates the annotation for the renamed molecules.
}
\examples{
data(data_normalized)
old_names <- rowData(data_normalized)$Molecule
# replace PCO with plasmenylPC
new_names <- sub("^LPE", "LysoPE", old_names)
update_molecule_names(data_normalized, old_names, new_names)
}
