% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidr-package.R
\name{use_interactive_graphics}
\alias{use_interactive_graphics}
\title{Activate interactive graphics}
\usage{
use_interactive_graphics(interactive = TRUE)
}
\arguments{
\item{interactive}{Should interactive plots be displayed? Default is TRUE.}
}
\value{
None
}
\description{
Use this function to turn on/off interactive graphics
plotting. Interactive plots require \code{plotly}
to be installed. Interactive graphics are disabled by default.
}
\examples{
data(data_normalized)
use_interactive_graphics()

# plot the variation in intensity and retention time of all measured
#  lipids in QC samples
d_qc <- data_normalized[, data_normalized$group == "QC"]
# plot_molecule_cv(d_qc, "Area")

# turn off interactivity
use_interactive_graphics(interactive = FALSE)
}
