% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfiles.R
\name{add_sample_annotation}
\alias{add_sample_annotation}
\title{Add sample annotation to Skyline data frame}
\usage{
add_sample_annotation(ds, annot_file)
}
\arguments{
\item{annot_file}{CSV file with at least 2 columns, sample names and group(s).}

\item{data}{Skyline data.frame created by \code{\link{read.skyline}}.}
}
\value{
Skyline data.frame with sample group information.
}
\description{
Add sample annotation to Skyline data frame
}
