% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_lipids}
\alias{annotate_lipids}
\title{Parse Molecule names to extract class and chain information.}
\usage{
annotate_lipids(data)
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link{read.skyline}}}
}
\value{

}
\description{
Parse Molecule names to extract class and chain information.
}
