% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{de_analysis}
\alias{de_analysis}
\title{Differential analysis of lipids between sample groups}
\usage{
de_analysis(..., data, measure = "Area", group_col = NULL)
}
\arguments{
\item{...}{expressions, or character strings which can be parsed to expressions, specifying contrasts. 
These are passed to \code{limma::makeContrasts}.}

\item{data}{Skyline data.frame created by \code{\link{read_skyline}}, should be normalized and log2 transformed}

\item{measure}{name of the column indicating sample names}

\item{group_col}{name of the column indicating sample groups}
}
\value{
TopTable as returned by limma package
}
\description{
Differential analysis of lipids between sample groups
}
\examples{
# type ?normalize_pqn to see how to normalize and log-transfome your data
data(data_normalized)
de_results = de_analysis(HighFat_water - NormalDiet_water, data=data_normalized, measure="Area")
}
