% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lsea}
\alias{lsea}
\title{Lipid set enrichment analysis (LSEA)}
\usage{
lsea(de.results, rank.by = c("logFC", "P.Value",
  "Adj.P.Val"))
}
\arguments{
\item{de.results}{Output of \code{\link[=de_analysis]{de_analysis()}}.}

\item{rank.by}{Statistic used to rank the lipid list.}
}
\value{
Enrichment results (data.frame) as returned from \code{\link[fgsea:fgsea]{fgsea::fgsea()}}.
The results also contain the following attributes:#' \itemize{
\item de.results    Original de.results input.
\item rank.by   Measure used to rank lipid molecules.
\item sets   Lipid sets tested, with their member molecules.
}
}
\description{
Lipid set enrichment analysis (LSEA)
}
\examples{
data(data_normalized)
de_results <- de_analysis(
  HighFat_water - NormalDiet_water,
  data = data_normalized, measure = "Area"
)
enrich_results <- lsea(de_results, rank.by = "logFC")
}
