% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{gen_lipidsets}
\alias{gen_lipidsets}
\title{Generate lipid sets}
\usage{
gen_lipidsets(data)
}
\arguments{
\item{data}{data.frame with a column named Molecule}
}
\value{
list of lipid sets
}
\description{
Generate lipid sets
}
