% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{gen_lipidsets}
\alias{gen_lipidsets}
\title{Generate lipid sets from lipid molecule names}
\usage{
gen_lipidsets(molecules)
}
\arguments{
\item{molecules}{A character vector containing lipid molecule names}
}
\value{
list of lipid sets
}
\description{
Generate lipid sets from lipid molecule names
}
\examples{
data(data_normalized)
molecules = rowData(data_normalized)$Molecule
gen_lipidsets(molecules)
}
