% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidr-package.R
\docType{data}
\name{lipidDefaults}
\alias{lipidDefaults}
\title{lipidDefaults.
A set of defualt mappings and annoataion used internally to correctly parse
lipid molecule names.}
\description{
lipidDefaults.
A set of defualt mappings and annoataion used internally to correctly parse
lipid molecule names.
}
\examples{
data(lipidDefaults)
}
