% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidr-package.R
\docType{data}
\name{lipidnames_pattern}
\alias{lipidnames_pattern}
\title{Patterns used in parsing lipid names}
\format{An object of class \code{list} of length 8.}
\usage{
lipidnames_pattern
}
\description{
A collection of patterns to extract lipid class and chain information
from lipid names. Used internally by the package.
}
\examples{
data(lipidnames_pattern)
}
\keyword{datasets}
