% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls.R
\name{mva}
\alias{mva}
\title{Plot a PCA plot to investigate sample clustering}
\usage{
mva(data, measure = "Area", method = c("PCA", "PCoA", "OPLS-DA"),
  group_col = NULL, groups = NULL, ...)
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link{read_skyline}}}

\item{measure}{which measure to use as intensity, usually Area_norm. The meausre should be already summarized and normalized}

\item{method}{either PCA, PCoA or OPLS-DA}

\item{group_col}{Sample annotation to use as grouping column}

\item{groups}{two groups to be used for supervised analysis (OPLS-DA), ignored in other methods.}

\item{...}{Extra arguments to be passed to \code{\link{opls}} for OPLS-DA, ignored in other methods.}
}
\description{
Blank samples are automatically detected (using TIC) and excluded. Missing data
are imputed using average lipid itensity across all samples.
}
\examples{
datadir = system.file("extdata", package="lipidr")
filelist = list.files(datadir, "data.csv", full.names = TRUE)
d = read_skyline(filelist)
clinical_file = system.file("extdata", "clin.csv", package="lipidr")
d = add_sample_annotation(d, clinical_file)

# PCA
mvaresults = mva(d, measure="Area", method="PCA")
plot_mva(mvaresults, color_by="group")

}
