% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_pqn}
\alias{normalize_pqn}
\title{Perform Probabilistic Quotient Normalization for intensities.}
\usage{
normalize_pqn(data, measure = "Area", exclude = "blank", log = TRUE)
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link{read_skyline}}}

\item{measure}{which measure to use as intensity, usually Area, Area.Normalized or Height}

\item{exclude}{Samples to exclude, can be either: \cr 
"blank" - automatically detected blank samples and exclude them
logical vector with the same length as samples}

\item{log}{whether the normalized values should be log2 transformed}
}
\description{
Perform Probabilistic Quotient Normalization for intensities.
}
\references{
Dieterle, F., Ross, A., Schlotterbeck, G., & Senn, H. (2006). 
Probabilistic quotient normalization as robust method to account for dilution 
of complex biological mixtures. Application in 1H NMR metabonomics. 
Analytical chemistry, 78(13), 4281-4290.
}
