% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_chain_distribution}
\alias{plot_chain_distribution}
\title{Plot a chart for log fold changes of lipids per class showing chain lengths and saturations}
\usage{
plot_chain_distribution(de_results, contrast = NULL, measure = "logFC")
}
\arguments{
\item{de_results}{output of \code{\link{de.analysis}}}

\item{contrast}{which comparison to plot.}

\item{measure}{which measure to plot the distribution of: logFC, P.Value, Adj.P.Val}
}
\description{
Plot a chart for log fold changes of lipids per class showing chain lengths and saturations
}
