% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_class_boxplot}
\alias{plot_class_boxplot}
\title{Plot a boxplot chart to examine distribution of values per class}
\usage{
plot_class_boxplot(data, measure = "Area", log = TRUE)
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link{read_skyline}}}

\item{measure}{which measure to plot the distribution of: usually Area, Area.Normalized or Height}

\item{log}{whether values should be log2 transformed}
}
\description{
The function is usually used to look at intensity distribution in each class,
but can also be used to look at different measures, such as  `Retention.Time` or `Background`.
}
\examples{
data(data_normalized)

plot_class_boxplot(data_normalized, "Area", log=TRUE)
plot_class_boxplot(data_normalized, "Retention.Time", log=FALSE)
}
