% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_class_enrichment}
\alias{plot_class_enrichment}
\title{Plot a boxplot chart to log fold changes of lipids per class}
\usage{
plot_class_enrichment(de_results, significant.sets, measure = "logFC")
}
\arguments{
\item{de_results}{output of \code{\link{de_analysis}}}

\item{significant.sets}{list of significantly changed lipid sets (output of \code{\link{significant_lipidsets}})}

\item{measure}{which measure to plot the distribution of: logFC, P.Value, Adj.P.Val}
}
\description{
The function is usually used to look at logFC distribution of lipids in each class,
marking significantly enriched classes. Can also be used to plot `P.Value` or `Adj.P.Val`
}
