% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_class_sd}
\alias{plot_class_sd}
\title{Plot a bar chart for standard deviation of a certain measure in each class}
\usage{
plot_class_sd(data, measure = "Area", log = TRUE)
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link[=read_skyline]{read_skyline()}}.}

\item{measure}{Which measure to plot the distribution of: usually Area,
Area.Normalized, Height or Retention.Time}

\item{log}{Whether values should be log2 transformed
(Set FALSE for retention time).}
}
\value{
A ggplot object.
}
\description{
The function is usually used to look at standard deviations of intensity in
each class, but can also be used to look at different measures such as
\code{Retention.Time}, to ensure all lipids are eluted within the expected range.
To assess instrumental variation apply the function to technical quality
control samples.
}
\examples{
data(data_normalized)

plot_class_sd(data_normalized, "Area", log = TRUE)
plot_class_sd(data_normalized, "Retention.Time", log = FALSE)
}
