% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls.R
\name{plot_mva}
\alias{plot_mva}
\title{Plot a multivariate scatterplot of sample scores to investigate sample clustering}
\usage{
plot_mva(mvaresults, components = c(1, 2), color_by = NULL)
}
\arguments{
\item{mvaresults}{Results obtained from \code{\link{mva}}.}

\item{components}{Which components to plot. Ignored for PCoA, OPLS and OPLS-DA results.}

\item{color_by}{Sample annotation to use as color.}
}
\value{
A ggplot of the sample scores.
}
\description{
Plot a multivariate scatterplot of sample scores to investigate sample clustering
}
\examples{
data(data_normalized)

# PCA
mvaresults <- mva(data_normalized, measure = "Area", method = "PCA")
plot_mva(mvaresults, color_by = "group")
plot_mva(mvaresults, color_by = "Diet", components = c(2, 3))

# PCoA
mvaresults <- mva(data_normalized, measure = "Area", method = "PCoA")
plot_mva(mvaresults, color_by = "group")

# OPLS-DA
mvaresults <- mva(data_normalized, method = "OPLS-DA", group_col = "BileAcid", groups = c("water", "DCA"))
plot_mva(mvaresults, color_by = "group")
}
