% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls.R
\name{plot_mva_loadings}
\alias{plot_mva_loadings}
\title{Title}
\usage{
plot_mva_loadings(mvaresults, components = c(1, 2), color_by = NULL,
  top.n = nrow(mvaresults$loadings))
}
\arguments{
\item{mvaresults}{Results obtained from \code{\link{mva}}}

\item{components}{which components to plot. Ignored for PCoA, OPLS-DA results}

\item{color_by}{Sample annotation to use as color}

\item{top.n}{Number of top ranked features to highlight on the plot}
}
\value{
A plot
}
\description{
Title
}
