% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_results_volcano}
\alias{plot_results_volcano}
\title{Plot a volcano chart for differential analysis results.}
\usage{
plot_results_volcano(de_results, show.labels = TRUE)
}
\arguments{
\item{de_results}{output of \code{\link{de_analysis}}}
}
\description{
Plot a volcano chart for differential analysis results.
}
