% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sample_tic}
\alias{plot_sample_tic}
\title{Plot a bar chart for total sample intensity}
\usage{
plot_sample_tic(data, measure = "Area", log = TRUE)
}
\arguments{
\item{data}{Skyline data.frame created by \code{\link{read.skyline}}}

\item{measure}{which measure to use as intensity, usually Area, Area.Normalized or Height}

\item{log}{whether values should be log2 transformed}
}
\description{
Plot a bar chart for total sample intensity
}
\examples{
datadir = system.file("extdata", package="lipidr")
filelist = list.files(datadir, "data.csv", full.names = TRUE)
d = read_skyline(filelist)

plot_sample_tic(d, "Area", log=TRUE)
plot_sample_tic(d, "Background", log=FALSE)
}
