% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfiles.R
\name{read_skyline}
\alias{read_skyline}
\title{Read Skyline exported files}
\usage{
read_skyline(files)
}
\arguments{
\item{files}{list of Skyline exported files in CSV format}
}
\value{
SummarizedExperiment object.
}
\description{
Read Skyline exported files
}
\examples{
datadir = system.file("extdata", package="lipidr")
filelist = list.files(datadir, "data.csv", full.names = TRUE) # all csv files
d = read_skyline(filelist)

View automatically generated lipid annotations
rowData(d)
}
