% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{significant_lipidsets}
\alias{significant_lipidsets}
\title{Get a list of significantly changed lipid sets}
\usage{
significant_lipidsets(enrich.results, p.cutoff = 0.05, size.cutoff = 2)
}
\arguments{
\item{enrich.results}{output of \code{\link{enrich.lipid.set}}}

\item{p.cutoff}{}

\item{size.cutoff}{}
}
\description{
Get a list of significantly changed lipid sets
}
