% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{significant_lipidsets}
\alias{significant_lipidsets}
\title{Get a list of significantly changed lipid sets}
\usage{
significant_lipidsets(enrich.results, p.cutoff = 0.05, size.cutoff = 2)
}
\arguments{
\item{enrich.results}{output of \code{\link{enrich_lipidsets}}}

\item{p.cutoff}{Sigificance threshold}

\item{size.cutoff}{Minimum number of lipids in a set tested for enrichment}
}
\description{
Get a list of significantly changed lipid sets
}
