% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{significant_lipidsets}
\alias{significant_lipidsets}
\title{Get a list of significantly changed lipid sets}
\usage{
significant_lipidsets(enrich.results, p.cutoff = 0.05, size.cutoff = 2)
}
\arguments{
\item{enrich.results}{Output of \code{\link{enrich_lipidsets}}.}

\item{p.cutoff}{Significance threshold.}

\item{size.cutoff}{Minimum number of lipids in a set tested for enrichment.}
}
\value{
A list of character vectors of significantly enriched sets for each contrast.
}
\description{
Get a list of significantly changed lipid sets
}
\examples{
data(data_normalized)
de_results <- de_analysis(HighFat_water - NormalDiet_water, data = data_normalized, measure = "Area")
enrich_results <- enrich_lipidsets(de_results, rank.by = "logFC")
significant_lipidsets(enrich_results)
}
