% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{SkylineExperiment}
\alias{SkylineExperiment}
\title{Constructor for Skyline experiment from list of assays}
\usage{
SkylineExperiment(assay_list, metadata, colData = NULL, rowData = NULL)
}
\arguments{
\item{assay_list}{A list or SimpleList of matrix-like elements,
or a matrix-like object. Passed to \code{\link[=SummarizedExperiment]{SummarizedExperiment()}}.}

\item{colData}{An optional DataFrame describing the samples (contains
clinical information). Row names, if present, become the column names of
the SkylineExperiment.}

\item{rowData}{A DataFrame object describing the rows (contains generated
lipid annotations). Row names, if present, become the row names of the
SummarizedExperiment object. The number of rows of the DataFrame
must be equal to the number of rows of the matrices in assays.}
}
\value{
SkylineExperiment object
}
\description{
Constructor for Skyline experiment from list of assays
}
