% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls.R
\name{topImportantLipids}
\alias{topImportantLipids}
\title{Extract top lipids from OPLS-DA results}
\usage{
topImportantLipids(mvaresults, top.n = 10)
}
\arguments{
\item{mvaresults}{Results obtained from \code{\link{mva}}.}

\item{top.n}{Number of lipids to return.}
}
\value{
A dataframe of \code{top.n} lipids with their annotations.
}
\description{
Extract top lipids from OPLS-DA results
}
\examples{
data(data_normalized)

mvaresults = mva(data_normalized, method = "OPLS-DA", group_col = "BileAcid", groups=c("water", "DCA"))
topImportantLipids(mvaresults, top.n=10)
}
